fixed errors in the port to the c bindings
altering the kernel launches in CloverCL.C as an error when creating the local and gloabl arrsy
generate chunk one of the values was set to double instrad of int
in upadatehalo had to change the enqueu macro so that err was set correctly

code now workds and gives the corect answers on the small long problem

+++ generate_chunk_kernel_ocl.C
@@ -180,7 +180,7 @@ void generate_chunk_kernel_ocl_(
     //        NULL,
     //        &event9);
 
-    err = clEnqueueWriteBuffer(CloverCL::queue_c, CloverCL::state_geometry_buffer_c, CL_FALSE, 0, *nm_stes*sizeof(double), state_geometry, 0, NULL, NULL); 
+    err = clEnqueueWriteBuffer(CloverCL::queue_c, CloverCL::state_geometry_buffer_c, CL_FALSE, 0, *nm_stes*sizeof(int), state_geometry, 0, NULL, NULL); 
     //CloverCL::queue.enqueueWriteBuffer(
     //        CloverCL::state_geometry_buffer,
     //        CL_FALSE,
